moved APIFeatures class into utils folder
