+++ controllers/departureController.js
@@ -1,63 +1,5 @@
 const Departure = require('../models/departureModel');
-
-class APIFeatures {
-  constructor(query, queryString) {
-    this.query = query;
-    this.queryString = queryString;
-  }
-
-  filter() {
-    // FILTERING
-    const queryObj = { ...this.queryString };
-    const excludedFields = ['page', 'sort', 'limit', 'fields'];
-
-    excludedFields.forEach((el) => delete queryObj[el]);
-
-    // ADVANCED FILTERING
-    let queryStr = JSON.stringify(queryObj);
-    //regular expression
-    queryStr = queryStr.replace(/\b(gte|gt|lte|lt)\b/g, (match) => `$${match}`);
-    // console.log(JSON.parse(queryStr));
-
-    this.query = this.query.find(JSON.parse(queryStr));
-    // let query = Departure.find(JSON.parse(queryStr));
-    return this;
-  }
-
-  sort() {
-    //SORTING
-    if (this.queryString.sort) {
-      const sortBy = this.queryString.sort.split(',').join(' ');
-      this.query = this.query.sort(sortBy);
-    } else {
-      this.query = this.query.sort('-createdAt');
-    }
-
-    return this;
-  }
-
-  limitFields() {
-    //FIELD LIMITING
-    if (this.queryString.fields) {
-      const fields = this.queryString.fields.split(',').join(' ');
-      this.query = this.query.select(fields);
-    } else {
-      this.query = this.query.select('-__v');
-    }
-    return this;
-  }
-
-  paginate() {
-    //PAGINATION
-    const page = this.queryString.page * 1 || 1;
-    const limit = this.queryString.limit * 1 || 10;
-    const skip = (page - 1) * limit;
-
-    this.query = this.query.skip(skip).limit(limit);
-
-    return this;
-  }
-}
+const APIFeatures = require('../utils/apiFeatures');
 
 exports.getAllDepartures = async (req, res) => {
   try {
